Refactor methods into main auth class
